use np.float32 for examples and tests

+++ examples/scheme1/paired_english.py
@@ -1,5 +1,6 @@
 import os
 import time
+import numpy as np
 import spikeinterface as si
 import spikeinterface.preprocessing as spre
 import spikeinterface.comparison as sc
@@ -35,7 +36,7 @@ def main():
     timer = time.time()
 
     # lazy preprocessing
-    recording_filtered = spre.bandpass_filter(recording, freq_min=300, freq_max=6000, dtype=float)
+    recording_filtered = spre.bandpass_filter(recording, freq_min=300, freq_max=6000, dtype=np.float32)
     recording_preprocessed: si.BaseRecording = spre.whiten(recording_filtered)
 
     with TemporaryDirectory() as tmpdir:
